backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
qwen==0.0.9
pyjwt==2.8.0
python-multipart==0.0.6
qdrant-client==1.7.0
markdown==3.5.1
alembic==1.13.1